third_party/blink/web_tests/platform/generic/fast/shapes/shape-outside-floats/shape-outside-floats-ellipse-margin-right-expected.txt

The right edges of the three black squares should follow the outer ellipse boundary and each square should appear on a subsequent line.

X
X
X
PASS elementRect('s1').top is 0
PASS elementRect('s2').top is 20
PASS elementRect('s3').top is 40
PASS successfullyParsed is true

TEST COMPLETE
